+++ src/statusbar/statusBar.ts
@@ -4,12 +4,12 @@
 *--------------------------------------------------------------------------------------------*/
 
 import { StatusBarAlignment, StatusBarItem, window, workspace } from "vscode"; 
-import { Locators } from "../vscode-project-manager-core/src/autodetect/locators";
-import { ProjectStorage } from "../vscode-project-manager-core/src/storage";
+import { Locators } from "../autodetect/locators";
+import { ProjectStorage } from "../storage/storage";
 import { codicons } from "vscode-ext-codicons";
-import { isRemoteUri } from "../vscode-project-manager-core/src/utils/remote";
-import { getCodiconFromUri } from "../vscode-project-manager-core/src/icons";
-import { Project } from "../vscode-project-manager-core/src/project";
+import { isRemoteUri } from "../utils/remote";
+import { getCodiconFromUri } from "../utils/icons";
+import { Project } from "../core/project";
 
 let statusItem: StatusBarItem;